# Gestión_publicaciones_Parcial/backend_academic_management/docs/openapi.json
{
  "swagger": "2.0",
  "basePath": "/",
  "paths": {
...
    "/api/publications/": {
      "get": {
        "responses": {
          "401": {
            "description": "No autorizado"
          },
          "200": {
            "description": "Operaci\u00f3n exitosa",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Publicaci%C3%B3n"
              }
            }
          }
        },
        "summary": "Lista todas las publicaciones",
        "operationId": "listar_publicaciones",
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Publicaciones"
        ]
      },
      "post": {
        "responses": {
          "401": {
            "description": "No autorizado"
          },
          "400": {
            "description": "Datos inv\u00e1lidos"
          },
          "201": {
            "description": "Publicaci\u00f3n creada exitosamente"
          }
        },
        "summary": "Crea una nueva publicaci\u00f3n",
        "operationId": "crear_publicacion",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Publicaci%C3%B3n"
            }
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Publicaciones"
        ]
      }
    }
  },
  "info": {
    "title": "API de Gesti\u00f3n de Publicaciones Acad\u00e9micas",
    "version": "1.0",
    "description": "API para gestionar publicaciones acad\u00e9micas, proyectos y datos de investigadores"
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Introduce tu token con el formato: Bearer {token}"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "host": "localhost:5000"
}